move launcher to cluster subpackage
